HoD: Fix DSVEdit potentially reading past the end of the door list
HoD's engine gives no way to tell where a door list ends for sure so DSVEdit has to guess, but if the data immediately after the last door resembles a door (or is another room's door), it could previously result in DSVEdit reading more than it should.

This commit implements a hack where the room's layer list is used as an end marker for the door list, so there should no longer be any chance of reading past the end.